Split out TokenCheckbox
