+++ apps/fluent-tester/src/TestComponents/Checkbox/TokenCheckbox.windows.tsx
@@ -1,0 +1,3 @@
+export const SizeCheckbox = () => {
+  return null;
+};
